backend/requirements.txt
bcrypt==4.0.1
passlib[bcrypt]
fastapi
uvicorn
python-jose[cryptography]
python-multipart
python-dotenv
requests
PyMuPDF